project2_web_summarizer/pyproject.toml
[project]
name = "web-summarizer"
version = "0.1.0"
description = "网页内容总结 Bot —— Tool Use + 结构化输出练习项目"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
    "beautifulsoup4>=4.12.0",
]

[project.scripts]
websum = "summarizer.main:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
